Fix open machete file action in the case missing machete file - add discover action to the error notification
