+++ frontend/actions/src/main/java/com/virtuslab/gitmachete/frontend/actions/toolbar/OpenMacheteFileAction.java
@@ -4,11 +4,15 @@ import static com.virtuslab.gitmachete.frontend.resourcebundles.GitMacheteBundle
 
 import com.intellij.dvcs.DvcsUtil;
 import com.intellij.ide.actions.OpenFileAction;
+import com.intellij.notification.NotificationAction;
+import com.intellij.notification.NotificationType;
+import com.intellij.openapi.actionSystem.ActionManager;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.application.WriteAction;
 import com.intellij.openapi.vcs.VcsNotifier;
 import git4idea.GitUtil;
 import git4idea.config.GitVcsSettings;
+import git4idea.repo.GitRepository;
 import io.vavr.control.Try;
 import kr.pe.kwonnam.slf4jlambda.LambdaLogger;
 import lombok.CustomLog;
@@ -36,9 +40,9 @@ public class OpenMacheteFileAction extends BaseProjectDependentAction {
 
     // When selected Git repository is empty (due to e.g. unopened Git Machete tab)
     // an attempt to guess current repository based on presently opened file
-    val selectedGitRepo = getSelectedGitRepository(anActionEvent);
-    val repo = selectedGitRepo != null
-        ? selectedGitRepo
+    val gitRepository = getSelectedGitRepository(anActionEvent);
+    val repo = gitRepository != null
+        ? gitRepository
         : DvcsUtil.guessCurrentRepositoryQuick(project,
             GitUtil.getRepositoryManager(project),
             GitVcsSettings.getInstance(project).getRecentRootPath());
@@ -57,10 +61,18 @@ public class OpenMacheteFileAction extends BaseProjectDependentAction {
         .getOrNull();
 
     if (macheteFile == null) {
-      VcsNotifier.getInstance(project).notifyError(/* displayId */ null,
+      val errorWithDiscover = VcsNotifier.STANDARD_NOTIFICATION.createNotification(
           /* title */ getString("action.GitMachete.OpenMacheteFileAction.notification.title.machete-file-not-found"),
           /* message */ getString("action.GitMachete.OpenMacheteFileAction.notification.message.machete-file-not-found")
-              .fmt(gitDir.getPath()));
+              .fmt(gitDir.getPath()),
+          NotificationType.ERROR);
+      errorWithDiscover.addAction(NotificationAction
+          .createSimple(getString("action.GitMachete.DiscoverAction.GitMacheteToolbar.text"),
+              () -> ActionManager.getInstance().getAction(DiscoverAction.class.getSimpleName())
+                  .actionPerformed(anActionEvent)));
+
+      VcsNotifier.getInstance(project).notify(errorWithDiscover);
+
     } else {
       if (macheteFile.isDirectory()) {
         VcsNotifier.getInstance(project).notifyError(/* displayId */ null,
@@ -69,6 +81,21 @@ public class OpenMacheteFileAction extends BaseProjectDependentAction {
       }
 
       OpenFileAction.openFile(macheteFile, project);
+    }
+  }
+
+  @UIEffect
+  public static void openMacheteFile(GitRepository gitRepository) {
+    val project = gitRepository.getProject();
+    val file = gitRepository.getMacheteFile();
+    if (file != null) {
+      OpenFileAction.openFile(file, project);
+    } else {
+      VcsNotifier.getInstance(project).notifyError(
+          /* displayId */ null,
+          /* title */ getString("action.GitMachete.OpenMacheteFileAction.notification.title.machete-file-not-found"),
+          /* message */ getString("action.GitMachete.OpenMacheteFileAction.notification.message.machete-file-not-found")
+              .fmt(gitRepository.getRoot().getPath()));
     }
   }
 }
