+++ frontend/resourcebundles/src/main/resources/GitMacheteBundle.properties
@@ -397,7 +397,7 @@ action.GitMachete.OpenMacheteFileAction.description=Open machete file
 action.GitMachete.OpenMacheteFileAction.notification.title.cannot-open=Failed to open machete file
 action.GitMachete.OpenMacheteFileAction.notification.title.same-name-dir-exists=Cannot create file 'machete': Directory with the same name exists
 action.GitMachete.OpenMacheteFileAction.notification.title.machete-file-not-found=Machete file has not been found
-action.GitMachete.OpenMacheteFileAction.notification.message.machete-file-not-found=Could not find machete file under the specified root ({0})
+action.GitMachete.OpenMacheteFileAction.notification.message.machete-file-not-found=Could not find machete file under the specified root ({0}).
 
 
 action.GitMachete.OverrideForkPointOfCurrentAction.text=Git Machete: Override Fork Point of Current Branch\u2026
